+++ CHANGELOG.md
@@ -9,11 +9,17 @@
 # 2.1.3
 - Fixed bug on analytics integration
 
+# 2.1.3-rc1
+- Updated TiledeskWhenExpression and directive json_condition v2
+
 # 2.1.2
 - Fixed bug on AiPrompt action with vllm models
 
 # 2.1.1
 - Improved support for multiple vLLM servers
+
+# 2.1.1-rc19
+- Added TiledeskWhenExpression and directive json_condition v2
 
 # 2.1.0
 - Added support to analytics
@@ -123,6 +129,8 @@
 
 # 1.3.0
 - added: AI_ENDPOINT env var
+
+# 1.3.0-rc1
 - added: ability to get 'none' as bodytype in webresponse
 
 # 1.2.2
@@ -137,6 +145,19 @@
 - changed: refactoring of DIrWebRequestv2
 - bug-fixed: erro while parsing webrequestv2 body
 
+# 1.1.5-rc4
+- added: webhook action (same as intent one)
+
+# 1.1.5-rc3
+- bug-fixed: jsonBody parse error in web-request-v2
+
+# 1.1.5-rc2
+- bug-fixed: cannot set status of undefined reading res.status in DirAssistant
+
+# 1.1.5-rc1
+- changed: refactoring web-request-v2
+- bug-fixed: jsonBody parse error in web-request-v2
+
 # 1.1.4
 - bug-fixed: slit is undefined in TiledeskChatbotUtils
 
@@ -145,6 +166,12 @@
 
 # 1.1.2
 - bug-fixed: minor improvement
+
+# 1.0.27-rc3
+- added: fixToken function in TiledeskService utils class
+
+# 1.0.27-rc2
+- added: specchToText function to transcript audio file
 
 # 1.0.27-rc1
 - changed: context for gpt-40 and gpt-40-mini